contrib_tools/imaging/gnosis-forge-ffmpeg/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
python-magic==0.4.27
